hypervisor: move location of struct VmmPageTable
